requirements_doc.txt
# requirements for doc
# no need to manul install it. (See doc/Makefile)
jinja2<=3.03
pandoc
sphinx
numpydoc
sphinx-rtd-theme
nbsphinx
jupyter-sphinx
sphinxcontrib-programoutput
xradarsat2
xarray-safe-rcm
xarray-safe-s1